// src/conversation/types/conversation-state.type.ts
export enum ConversationState {

  IDLE = 'IDLE',

  GREETING = 'GREETING',

  SEARCHING = 'SEARCHING',

  WAITING_PRODUCT = 'WAITING_PRODUCT',

  WAITING_VEHICLE = 'WAITING_VEHICLE',

  WAITING_STOCK_CONFIRMATION = 'WAITING_STOCK_CONFIRMATION',

  TRANSFER_HUMAN = 'TRANSFER_HUMAN',

  FINISHED = 'FINISHED',
}